Able to save endo imgs through out use
I copied the way the endo imgs are saved and shown from the "Imaging" case and copy pasted them it in the "Image Prep Screen".

It will always try to take and save those endo imgs now. However, it is currently broken. It is stuck in a infinite while loop. To fix this you have to change the case that the software switches to after initialize. Change it to imaging. You also have to add a button property node for the pause button. Enable it. This way the loop can be broken and things will start to work like normal...

Changes actually need to be added so we don't have this janky way of doing it.