Changes that allow the video to save endo frames
I don't actually know if this is the right file. I ended up forgetting to switch to the test version and did all my work on the main repo.

Either way, I was able to change some stuff so that we can save endo frames at all times. There is another state created called save enfo frames or something like that.

We move here at the start of the software. And can only change again after the calibratio button or play buttton are clicked. It checks for the changes through something called channel. Idk what they is.

I don't want to type anymore. So I'm gonna let it go and just push this hoe.